Make path comparison case insensitive
On one mac os x system, the directory is reported as /Users/username/exercism in one place
and /users/username/exercism in another. Since both mac and windows are case insensitive in terms
of creating directories on a system (cannot have Hello and hello in the same location), lowercasing
both the filename and the directory before comparing them seems like a fairly innocuous approach.