Simplify config package
We only ever do three things from the outside:

1. Load an existing config, falling back to defaults if it doesn't exist.
2. Update values on a config.
3. Write the config to disc.

Everything else is details.

In order to be able to set a fake homeDir on this, the New function
delegates to a separate load function. This also allows us to avoid
potential race conditions in the tests by not setting ENV variables.