Report 'new' problems last after fetching.
Prior to this the 'unsubmitted' problems were listed at the bottom, and
that list is always going to be huge, including all the languages
someone doesn't care about. This means that their terminal window might
very well be filled with unsubmitted exercises, and they might not see
the newly fetched problem.

Putting the new problem at the bottom means that it's closest to their
PS1 and therefore more likely to be noticed.