Fixes #2303: cli cannot read file contents unless it's ascii or utf-8
Fixes issue #2303 in the exercism.io project.
https://github.com/exercism/exercism.io/issues/2303

Attempt to detect the charset of a submission so that it can be properly
transcoded to UTF-8.

Added UTF-16 and UTF-16BE fixture files for the test suite.

Changed the tests slightly to deal with different path and newline
separators on Windows.

Update travis config to include new dependencies.