Add a guard to the unsubmit command.
The unsubmit command does not take any arguments, but it would be
easy to assume that it does, considering how most of the other commands
work.

This changes the command so that it stops with a warning if someone
tries to pass an argument.

This change is in response to https://github.com/exercism/exercism.io/issues/2446
where someone tried unsubmitting a submission by passing the file path
to the unsubmit command, which blithely unsubmitted the most recent
submission--period.

It's confusing and unfriendly.