Adding `exercism status $language` command
This commit adds the functionality for the new command `exercism status $language`. The API endpoint has yet to be implemented, but I will work on that once this is deemed finished. I've tested this out with an API on my local machine serving up mock data that will match the output of the live API once I've implemented it, and it is working as expected.