Clarify trackID/language distinction on Problem type
The API is currently returning the track ID twice, once for track ID and
once for language. This fixes the CLI code so that if we fix the API,
the CLI will not break.

However, we cannot fix the API until we have released the next version
of the CLI, since doing so would break on existing versions.