Ditch the demo command
The demo was great when we had two or three languages, but now that we have almost
30 it makes no sense to download the first problem in each track.

The website no longer shows the demo command, but rather encourages people
to download the first problem in a given track explicitly.

We will also be changing the fetch command to not fetch things in unstarted
tracks, which is even more reason to not have someone implicitly start all
the tracks.