Changes `exercism fetch` behavior to remove filesystem noise
This commit changes the behavior of `exercism fetch` when called without any arguments. Instead of fetching and writing exercises for all tracks, it will now fetch and update only for tracks the user has folders for on their local environment. If a user calls `exercism fetch` without having any language tracks in their exercises folder, a prompt is shown to the user to tell them how to list all available tracks and how to fetch their first track for a language of their choice.