Add extra hints to submit on first iteration
As per #77 I've added the "extra hints" from the "beer song" exercise
README to the output of the submit command. These hints provide
direction to help refactor, and then participate in the community
through commenting on the exercism site.

To keep annoyance down and keep users from getting used to (and thus
ignoring) them, the extra hints will show only on a successful
submission of the first iteration of each exercise.