Download: make files with mode 0644 rather than 0755
When downloading a submission with `exercism download`, previously the
files would get written with 0755 (u+rwx g+rx o+rx). In discussion in
issue #276 we decided we would like the files to be non-executable, as
we don't want to encourage the behavior of executing an arbitrary remote
file.

The user still has the option of either running the file in the
appropriate manner for the language or explicitly chmodding if desired.

Fixes #276