Remove random encouragement after submitting
The post-submit message turned out to not work. After seeing it
the first time it just becomes noise, and you end up hunting for the
important information.

This simplifies the message, removing unnecessary noise. The only
information necessary is: language track, problem name, and the
URL of your iteration.

I put the url of the iteration alone on its own line instead of saying
"your submission can be found online at URL", since the extra verbiage
is superfluous, and it also prevents double-clicking on the line to
highlight the url and paste it easily into the browser.