The url for related solutions are not correct.
When user submit an exercise, the url of related solutions will be printed
by the CLI.  The printed url missed one `/` which makes it not accessible.

Here is the error output:

```
See related solutions and get involved here:
http://exercism.iotracks/cpp/exercises/binary
```

The fix is to add the missing `/` between `exercism.io` and `tracks`.