iteration_test: Accept suffixes in addition to prefixes
An upcoming commit will soon add a test file for which the suffix is
important, in addition to the prefix. The middle of the file will
contain long text that is not desirable to test. Thus, testing suffixes
is helpful.

Note that all current tests have no suffixes, so all cases should pass.