Add separate user config
This is the very first tiny step in a larger refactoring, just enough to
be able to refactor the debug command.

There are three distinct types of configuration:

- user configuration about their workspace and communicating with the API
- api configuration (hosts, endpoints)
- CLI configuration (messages, error strings, templates for
  communicating with the user)

I'd like to make these live in the {user,api,cli} packages
respectively.

It's taken me a long time to figure out what to call the directory tree
where people work on exercises. We could call it the exercism directory,
but that is ambiguous, because the configuration directory is going to
be '$CONFIG_HOME/exercism'. Also, if you're a maintainer, you might have
an exercise directory that contains all the various codebases and things
for working on the project.

I finally thought of 'workspace' which feels really clear.