Normalize the init function name in configure
Even though we're initializing configuration values here, the general concept is that
we're doing custom initialization for this particular command. In other commands we
will need to initialize flags but not config.

To make this a bit clearer I'm going with initSomethingCmd where Something is the name of
the command.