Figure out where to download an exercise to
Nextercism is going to let people have separate solutions exploring the same problem.
Each solution can contain any number of iterations. So someone might be exploring an
object oriented approach and a functional approach to the same exercise, at the same time.

They will be able to do this in directories that are numbered serially, so the first one
would be 'thingy', then the second would be 'thingy-2'.

When downloading an exercise, this means that we need to figure out which of the existing
directories for the exercise of the same name is the one that we want to write to.

If it's none of them, then we need to make a new directory.