Detect the type of a path
We are pretty flexible in the type of argument that we receive. It might be
an actual path, or it might be a conceptual path: the name of an exercise that
we need to walk the workspace to find. If it's a conceptual path we might find
more than one, in which case we'll do an interactive thing to disambiguate. If
it's a real path, then we can do a lot less guessing.