Create small abstraction for interactive -uh- interactions
If someone tries downloading or submitting something before
they've configured their token or workspace, we're going to want
to ask them to paste it to us.

For the workspace we want to be able to provide a default suggestion.

This Question abstraction provides a very small wrapper around such
an interaction, and makes it testable by providing the Reader/Writer
which can be overwritten with pre-populated readers, or a writer that
discards everything.