Remove optional unmarshal in client.Do()
If we unmarshal here we have to handle error messages,
but the error messages will need to be customized based
on the user's configuration, and also based on
error types that the API returns.

The api package should be a very thin wrapper around http.
It shouldn't have to know about the implementation details
of the API error messages and CLI error handling.