Tweak config track test for readability
When naming a method that contains a small conditional, I find it easier
to name if the method contains both the exception and the rule.

The notIfNeeded method is naming only the case where it's not needed.
This changes the logic so that it contains two variations on the same
concept ("how acceptable is this").