Add indirection to configure command
This splits up the configure command so that we can inject
some configuration values.

Right now there's a lot of pain in test setup where we need
to put together entire cobra commands and temporary directories
and environment variables.

This is slowly moving us in the direction of being able to inject
defaults and therefore test the command without manipulating
the environment so heavily.