Fix token validation logic
The configure command was calling the token validation endpoint, but
was not actually bailing or providing an error in the case of an invalid
token.

We will need to improve the error message to include the link to the
url where the token can be found. Also, we should probably not print
the current configuration after printing the error message.