Rename skip-auth flag on configure command
We need to verify both the API base URL and the token.
The name 'skip auth' is only applicable to the token.
To not be specific to auth, I went with the idea of 'verification'.

I chose 'no-verify' as the flag name, as this will be familiar to
people who use git's no-verify flag to skip hooks.