Delete redundant tests for configure command
The original tests go all the way through the top of the cobra command
and use the file system.

The logic remaining in the RunE function itself is now minimal, and all
the complexity has been isolated ot the runConfigure function, which
we've now decoupled enough to test it without the filesystem.