Use sensible defaults for Workspace dir
We are phasing out the UserConfig type, but until we
do, we are hooking into the new sensible os-specific
path default for workspace.

This moves the "do not clobber" logic into the user config
type's SetDefaults method. As we overhaul the verification
logic in the 'configure' command, we will likely move that
out of config and into cmd.