Add indirection to the submit command
The test setup/teardown is quite arduous right now.
This splits apart the submit command so that we can set up
all the environment and configuration in the RunE function, and
then pass the configuration value along with flags and arguments
to the runSubmit function.

This lets us ignore the cobra.Command setup/teardown in the command
tests.