Make CLIConfig loading explicit in submit command
We're working to get rid of the embedded viper config pattern
that we currently use for the UserConfig and CLIConfig types, as
these rely on a lot of implicit context which makes testing much
more difficult.

We still have an explicit Load() in the runSubmit command which we
will need to get rid of, but this will require clarifying the behavior
around the track preparation. We should be able to do this in
such a way that we don't have to load a new configuration off the
filesystem.