Update submit command to take files as arguments
We decided to restructure the UI for the submit command (again).

   exercism submit file1 file2 file3

When we re-introduce the ability to submit a directory we will do

   exercism submit dir

Then we check if there are multiple arguments.  If there are, then
they're submitting files and we fail for all the usual reasons. If there
is just one argument, then we check first if it's a file or a directory
and deal with it appropriately.

But for now: just files.