Handle submitting relative paths to files
To test relative paths we change directory into
the temporary directory and then pass the relative
directory as an argument.

On the Mac, temporary directories are created within /var
which is a symlinked directory under /private:

    $ ls -l /var
    lrwxr-xr-x@ 1 root  wheel  11 Dec 17  2017 /var -> private/var

Temporary directories are returned using the path within the symlinked
path (/var/folders/...) rather than the source path (/private/var/folders/...).

To test the relative path change directory into the temporary directory,
then pass the relative path as an argument, which then gets resolved
into an absolute path.

Before submitting, we make sure that the file is within the configured
workspace. However, this now causes a mismatch, because the
workspace path is the one with the symlink, whereas the resolved
absolute path in this case is the source path.

To fix this, we evaluate the symlinks in the workspace path
and resolve these before comparing them to the submitted files.

The Windows build fails with this test, so I've separated the relevant
tests out into a conditional build. Once we get it fixed it may make
more sense to inline it into the main submit test and do a little bit of
conditional setup.