Get rid of transmission type
This inlines the little bit of logic that we needed from the workspace.Transition
and deletes the type.

We may find that we want to re-introduce an abstraction for some of this,
but I want to write it all down as simply as possible first and then go from there.