Rename test helper for clarity and consistency
The various command tests need different types of mock servers. These all live in the
same namespace, despite being defined in different files.

The download command test had a 'makeMockServer' function, which in the scope of
the entire cmd package is not specific enough. This renames it to fakeDownloadServer
which follows the pattern used in the submit test (fakeSubmitServer).

I prefer that the name reflect _what it returns_ rather than _what you
are asking it to do_. So for something that returns a _thing_, I'd
rather it be named _thing_ than _make thing_.