Complain if downloading without a configured token
We need to know who you are in order to download the correct exercise with all the requisite
metadata.

Originally the download command allowed you to pass a token on the fly, during download.
That added a lot of complexity, so instead we bail with a helpful error message
that explains how to configure the client.