Use config dir, not file, in troubleshoot command
In the old CLI we had a single config file. Now we potentially have several.
They will all live in the config directory.

The troubleshoot command now outputs the directory instead of the file.

If the troubleshoot command is unconfigured, this would previously have
given an empty string for the config file. Now it will always print the
default config directory, unless the user has specifically defined
an override using environment variables.