Delete speculative implementation of prepare command
When I first started writing the new CLI I had some grand ideas about how submit should work,
which required all sorts of complicated guessing, which in turn required a command to
store settings on a per-track basis.

We may still need to prepare things on a per-track basis, but the current implementation was
highly speculative, and I decided that it would be better to delete it and start fresh
when we know exactly what we need.