Tweak output of troubleshoot command when the CLI is unconfigured (#675)
* Pass full configuration value to troubleshooting status

Instead of passing just the viper config, which could be empty,
pass the full config.Configuration value, which has access to defaults.
This will let us give better debugging output.

* Fix output for 'home' in troubleshoot command

The home directory is never written to the viper config, we get it from the environment,
which is stored in the config.Configuration value.

* Output default workspace in troubleshoot if unconfigured

Rather than providing an empty string for the workspace in the
troubleshoot command when the CLI is unconfigured, this now says

    Workspace: /the/actual/path (default)

* Use config dir, not file, in troubleshoot command

In the old CLI we had a single config file. Now we potentially have several.
They will all live in the config directory.

The troubleshoot command now outputs the directory instead of the file.

If the troubleshoot command is unconfigured, this would previously have
given an empty string for the config file. Now it will always print the
default config directory, unless the user has specifically defined
an override using environment variables.

* Inline token in troubleshoot config status

It was unnecessary to have an additional local variable.

* Use proper helper method for settings URL in troubleshoot command

The config package has logic for the settings URL, we shouldn't be hard-coding it here.

* Fix ping when troubleshoot is unconfigured