Rename solution to metadata
What we previously called Solution is actually the exercise metadata, which also
includes the user's solution UUID.

In order to keep things as clear as possible, this also renames the Solutions
type to MetadataCollection. I couldn't think of anything better than collection to deal with the
multiples of a multiple.