Get rid of the workspace.Locate behavior and comms package (#696)
* Simplify open command

This no longer accepts just an exercise name. It mirrors the behavior of
the submit command where you pass the path to the exercise you want to
open in the browser.

In the future we can expand this to take --exercise, --track, and --team.

* Simplify submit command

We determine the exact exercise directory based on the filepaths.
Since we have the directory, we don't need to do complicated guessing,
we can just load in the exercise metadata directly.

* Rename slug to param in download command for clarity

* Rename exercise to slug in download for clarity

* Rename dir variable in download command for clarity

We were overloading the variable, which makes it confusing.

* Add missing error handling in download command

* Add metadata dir method to exercise type

We're going to want to ensure that the directory exists.
For the moment, this directory is the same as the exercise directory,
but we're working towards putting this in a subdirectory to match
common industry conventions.

* Simplify download command to rely on Exercise type

* Get rid of solution path abstraction in workspace

We've moved the idea of the exercise metadata path onto the Exercise type.

* Delete complicated Locate behavior in workspace

It's no longer used.

* Delete unused comms package

* Tweak download command for clarity

Simplify a conditional.