Fix windows filepaths that accidentally got submitted to the server
Some of the older CLIs did not submit a normalized path for the file, but rather submitted the
filepath. This means that we sometimes had forward slashes and sometimes had backslashes.

The CLI now always submits the normalized path with forward slashes.

This ensures that when we receive files submitted with a buggy client we rewrite the paths
correctly.