Handle exercise directories with numeric suffixes
An early design decision was to allow people to work on multiple solutions to the same
exercise at the same time. We later decided not to do this, but the CLI still had
logic that supported it.

This caused us to (incorrectly) create exercise directories that were named with a
numeric suffix, which would then get submitted to the backend server, which doesn't
care about filepaths.

Because the exercise with the numeric suffix doesn't match the expected path, we did
not correctly trim off any leading directories on submit, which further caused
the download command to put the solution in a weirdly and deeply nested directory,
making the solution hard to find and review.