Wrap payload & config to avoid passing config
Previously getExercise & getMetadata required passing in a config which
is redundant because it's already required when creating a newDownloadPayload.

Since the config cannot be put directly on downloadPayload, these changes
resolve the problem by making a new struct that holds references to
config & downloadPayload.

An additional benefit of these changes is that the methods get moved
off downloadPayload and it no longer contains behavior.