exercise has tacit dep on filesBelongToSameExercise
exercise() uses ws.ExerciseDir() to find the root dir of an exercise.
Since ws.ExerciseDir requires a single path, this tacitly assumes that
the submitted paths are all part of the same exercise. There is
a validation for this called `filesBelongToSameExercise`. This isn't an
issue from the point of view of `runSubmit` but it might be an issue if
`exercise()` gets moved around in the future as this isn't explicitly
stated. These changes attempt to make it more clear by stating it in the
godoc.