Adapt downloadableFrom to use template pattern
This seems like an appropriate use of the template pattern. At present
it seems like we have a reasonable default to fallback on, but we know
that creating from flags is an invariant. Using the template pattern here
allows for future types to delegate to the default or add their own
specialization.

Taking advantage of embedding makes this more concise. Types that
delegate to the default just need to embed downloadParams. This gives
the flexibility to implement any number of methods without having to
define the entire interface.