Extract logic for the URI variable identifying the exercise to download
There is very likely a better abstraction hiding close by.
This is a minimal, cohesive change, which I think is fairly uncontroversial.

If we do find the better abstraction, it's likely that we'll be able to move this one
to it wholesale.

I've left a bit of duplication in place for now, rather than make a big
guess about the abstraction.