Allow to configure the workspace to be a symlink
Using a symlink for the workspace is already supported and, if created
after initializing the configuration, works without issues. configure
command however doesn't recognize symlinks and threat them the same way
as a file existing at the workspace location, bailing out.

Change the configure command to allow using a symlink for a workspace if
it points to an existing directory. If the symlink points to a file,
threat it the same way as an ordinary file at workspace path. If the
symlink is broken, give an explanatory message to the user.