goreleaser: add arm64 build for each OS
The latest release of the Exercism CLI (3.1.0, 2022-10-04) includes
these assets, via GoReleaser:

    exercism-3.1.0-darwin-x86_64.tar.gz
    exercism-3.1.0-freebsd-i386.tar.gz
    exercism-3.1.0-freebsd-x86_64.tar.gz
    exercism-3.1.0-linux-armv5.tar.gz
    exercism-3.1.0-linux-armv6.tar.gz
    exercism-3.1.0-linux-i386.tar.gz
    exercism-3.1.0-linux-ppc64.tar.gz
    exercism-3.1.0-linux-x86_64.tar.gz
    exercism-3.1.0-openbsd-i386.tar.gz
    exercism-3.1.0-openbsd-x86_64.tar.gz
    exercism-3.1.0-windows-armv5.zip
    exercism-3.1.0-windows-armv6.zip
    exercism-3.1.0-windows-i386.zip
    exercism-3.1.0-windows-x86_64.zip
    exercism_checksums.txt
    exercism_checksums.txt.sig

For future releases, this commit configures GoReleaser to include an
arm64 (also known as aarch64) release asset for each of the existing
OSes. Note that we will have separate x86_64 and arm64 release assets
for macOS; previously macOS users on arm64 had to run the x86_64 binary
via Rosetta (or build the CLI themselves).

At least for now, let's avoid adding a fat binary. But note that
GoReleaser does support it [1].

[1] https://goreleaser.com/customization/universalbinaries/

Closes: #966