go.mod, go.sum: bump `golang.org/x/sys` from 20170803 to 20201202
This commits the result of running

    go get golang.org/x/sys@v0.0.0-20201202213521-69691e467435

to fix a build failure for freebsd_arm64.

This seems to be the minimum bump to fix that failure without
introducing a build failure for darwin_arm64 and openbsd_arm64.

We cannot bump to the latest release (0.3.0, 2022-12-03) because that
causes the Linux and macOS CI jobs to fail with:

    Error: ../../../go/pkg/mod/golang.org/x/sys@v0.3.0/unix/syscall.go:83:16: undefined: unsafe.Slice
    Error: ../../../go/pkg/mod/golang.org/x/sys@v0.3.0/unix/syscall_darwin.go:95:8: undefined: unsafe.Slice
    Error: ../../../go/pkg/mod/golang.org/x/sys@v0.3.0/unix/syscall_unix.go:118:7: undefined: unsafe.Slice
    Error: ../../../go/pkg/mod/golang.org/x/sys@v0.3.0/unix/sysvshm_unix.go:33:7: undefined: unsafe.Slice
    note: module requires Go 1.17

and we need other changes in order to bump the minimum Go version.