goreleaser: add arm64 build for each OS (#1073)
* goreleaser: add arm64 build for each OS

The latest release of the Exercism CLI (3.1.0, 2022-10-04) includes
these assets, via GoReleaser:

    exercism-3.1.0-darwin-x86_64.tar.gz
    exercism-3.1.0-freebsd-i386.tar.gz
    exercism-3.1.0-freebsd-x86_64.tar.gz
    exercism-3.1.0-linux-armv5.tar.gz
    exercism-3.1.0-linux-armv6.tar.gz
    exercism-3.1.0-linux-i386.tar.gz
    exercism-3.1.0-linux-ppc64.tar.gz
    exercism-3.1.0-linux-x86_64.tar.gz
    exercism-3.1.0-openbsd-i386.tar.gz
    exercism-3.1.0-openbsd-x86_64.tar.gz
    exercism-3.1.0-windows-armv5.zip
    exercism-3.1.0-windows-armv6.zip
    exercism-3.1.0-windows-i386.zip
    exercism-3.1.0-windows-x86_64.zip
    exercism_checksums.txt
    exercism_checksums.txt.sig

For future releases, this commit configures GoReleaser to include an
arm64 (also known as aarch64) release asset for each of the existing
OSes. Note that we will have separate x86_64 and arm64 release assets
for macOS; previously macOS users on arm64 had to run the x86_64 binary
via Rosetta (or build the CLI themselves).

At least for now, let's avoid adding a fat binary. But note that
GoReleaser does support it [1].

[1] https://goreleaser.com/customization/universalbinaries/

Closes: #966

* go.mod, go.sum: bump `golang.org/x/sys` from 20170803 to 20201202

This commits the result of running

    go get golang.org/x/sys@v0.0.0-20201202213521-69691e467435

to fix a build failure for freebsd_arm64.

This seems to be the minimum bump to fix that failure without
introducing a build failure for darwin_arm64 and openbsd_arm64.

We cannot bump to the latest release (0.3.0, 2022-12-03) because that
causes the Linux and macOS CI jobs to fail with:

    Error: ../../../go/pkg/mod/golang.org/x/sys@v0.3.0/unix/syscall.go:83:16: undefined: unsafe.Slice
    Error: ../../../go/pkg/mod/golang.org/x/sys@v0.3.0/unix/syscall_darwin.go:95:8: undefined: unsafe.Slice
    Error: ../../../go/pkg/mod/golang.org/x/sys@v0.3.0/unix/syscall_unix.go:118:7: undefined: unsafe.Slice
    Error: ../../../go/pkg/mod/golang.org/x/sys@v0.3.0/unix/sysvshm_unix.go:33:7: undefined: unsafe.Slice
    note: module requires Go 1.17

and we need other changes in order to bump the minimum Go version.