github/workflow/release.yml: Add automated release pipeline
This change introduces a GitHub action workflow for publishing release
artifacts generated by goreleaser when a new tag is pushed.

* Add install walk-through to release notes header
A markdown file under a hidden release directory was added for use
with the gorleaser configuration for prepending a static header to the goreleaser generated
changelog.

* Update goreleaser configuration to generate installer archives
The Exercism installer requires the Windows archives to be
named slightly differently. In order to generate the archives as part
of the release process goreleaser has been split into two builds:
release-build and installer-build. The two builds are then archived
separately using the name_template configuration to generate the
properly named zip files. All generated archives are included in the final
release checksum.

* release.md: Updates release documentation to reflect the steps used by
  the automated pipeline